added better version message

+++ libs/ma/src/util/support.cpp
@@ -15,6 +15,7 @@
 #include <sys/stat.h>
 #include <sys/types.h>
 #include <vector>
+#include <cctype>
 
 #include "util/exception.h"
 #include "util/support.h"
@@ -37,8 +38,8 @@ bool EXPORTED fileExists( const std::string& rsFile )
 bool EXPORTED is_number( const std::string& s )
 {
     return !s.empty( ) &&
-           std::find_if( s.begin( ) + 1, s.end( ), []( char c ) { return !std::isdigit( c ); } ) == s.end( ) &&
-           ( std::isdigit( s[ 0 ] ) || s[ 0 ] == '-' );
+           std::find_if( s.begin( ) + 1, s.end( ), []( char c ) { return !std::isdigit( (int)c ); } ) == s.end( ) &&
+           ( std::isdigit( (int)s[ 0 ] ) || s[ 0 ] == '-' );
 } // function
 
 bool EXPORTED ends_with( const std::string& rsX, const std::string& rsEnd )
